pyproject.toml
[project]
name = "data-analysis-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.97.0",
    "polars>=0.20.0",
    "python-dotenv>=1.1.1",
]
